+++ web/src/main/java/org/springframework/security/web/PortMapperImpl.java
@@ -104,8 +104,8 @@ public class PortMapperImpl implements PortMapper {
 			Integer httpPort = Integer.valueOf(entry.getKey());
 			Integer httpsPort = Integer.valueOf(entry.getValue());
 
-			if ((httpPort.intValue() < 1) || (httpPort.intValue() > 65535)
-					|| (httpsPort.intValue() < 1) || (httpsPort.intValue() > 65535)) {
+			if ((httpPort < 1) || (httpPort > 65535)
+					|| (httpsPort < 1) || (httpsPort > 65535)) {
 				throw new IllegalArgumentException(
 						"one or both ports out of legal range: " + httpPort + ", "
 								+ httpsPort);
